Create DiskFormatTrack.cpp for class FormatTrack. Move log macros into new DiskLog.h
