+++ AppleWinExpress2008.vcproj
@@ -26,7 +26,6 @@
 			>
 			<Tool
 				Name="VCPreBuildEventTool"
-				Description=""
 				CommandLine=".\test\UnitTests.bat Release"
 			/>
 			<Tool
@@ -132,7 +131,6 @@
 			>
 			<Tool
 				Name="VCPreBuildEventTool"
-				Description=""
 				CommandLine=".\test\UnitTests.bat Debug"
 			/>
 			<Tool
@@ -835,6 +833,14 @@
 					>
 				</File>
 				<File
+					RelativePath=".\source\DiskFormatTrack.cpp"
+					>
+				</File>
+				<File
+					RelativePath=".\source\DiskFormatTrack.h"
+					>
+				</File>
+				<File
 					RelativePath=".\source\DiskImage.cpp"
 					>
 				</File>
@@ -848,6 +854,10 @@
 				</File>
 				<File
 					RelativePath=".\source\DiskImageHelper.h"
+					>
+				</File>
+				<File
+					RelativePath=".\source\DiskLog.h"
 					>
 				</File>
 				<File
